Rename gevent bootstrap package
